refactor(bench): support different quantizers
This will allow to compare bnb, awq, gptq and hqq more easily.
